Optimize YYError
- Optimized YYError calls from ~1 second to ~300 microseconds
  - The symbol table is built on YYTK init
  - Switched from using std::map to a vector for faster building without extra allocations
- Added engine stacktracing for YYError
- Fixed warning on casting from int64_t to size_t
- Updated Aurie headers

+++ YYToolkit/source/YYTK/Module Interface/MI_Public.cpp
@@ -682,7 +682,7 @@ namespace YYTK
 		if (!AurieSuccess(last_status))
 			return last_status;
 
-		Size = possible_size.ToInt64();
+		Size = static_cast<size_t>(possible_size.ToInt64());
 		return AURIE_SUCCESS;
 	}
